Providing instructions for fixing failed compression.
There is a bug in the phar extension that prevents archives with a large number of files from being compressed. For some reason, it either leaves file handles open or opens all of the files at once. This trips the maximum limit allowed by the system and causes an exception to be thrown. This tweak will make it less confusing and help developers find a workaround.